Add files via upload
The purpose is to build a classification model using the ’Predictive maintenance’ dataset, which consists of 10 000 data points stored as rows with 8 features in columns. The classifier will have to be able to predict the target variable, which takes value ’0’ if the machine has no failure and therefore no maintenance is needed, ’1’ if on the contrary, if some kind of damage has been revealed in the machine and it needs maintenance.
The dataset is taken from kaggle: https://www.kaggle.com/shivamb/machine-predictive-maintenance-classification.
To build such a model, the original dataset was split into 3 parts: the ’training set’ (60% of the data), the ’validation set’ (20% of the data) and the ’test set’ (20%
of the data). Our model will be built by the use of training and validation. In particular, the former will be used for us to construct three classification models
such as the ’logistic regression’, the ’random forest’ and a ’neural network’. The task of the second, on the other hand, will be to choose which of the three models
performs best. 
In the followi